Updated with docker-compose env file
