+++ scraping/the_star/.env
@@ -1,0 +1,23 @@
+
+# PRODUCTION
+IP=
+FLASK_ENV=production
+
+NETWORK=webproxy
+
+NGINX_WEB=nginx-web
+DOCKER_GEN=nginx-gen
+LETS_ENCRYPT=nginx-letsencrypt
+
+DOMAINS=
+LETSENCRYPT_EMAIL=edmund.hee05@gmail.com
+
+#
+## NGINX file path (mount into the host)
+# Here you can configure the path where nginx stores all the configurations and certificates.
+# With the value ./nginx-data it creates a new sub-folder into your current path.
+
+NGINX_FILES_PATH=./nginx-data
+
+DOCKER_HTTP=80
+DOCKER_HTTPS=443
