+++ scraping/the_star/Dockerfile
@@ -5,12 +5,14 @@ MAINTAINER LEAD "napster@thelead.io"
 ENV HOME=/home/coronatracker
 
 WORKDIR $HOME
-RUN mkdir $HOME/coronatracker
+RUN mkdir $HOME/app/
 
-COPY the_star.py $HOME/
-COPY db_connector.py $HOME/
-COPY db.json $HOME/
-COPY requirements.txt $HOME/
+COPY the_star.py $HOME/app/
+COPY db_connector.py $HOME/app/
+COPY utils.py $HOME/app/
+COPY __init__.py $HOME/app/
+COPY db.json $HOME/app/
+COPY requirements.txt $HOME/app/
 
-RUN python3.7 -m pip install -r $HOME/requirements.txt
-CMD [ "python" , "/home/coronatracker/the_star.py"]
+RUN python3.7 -m pip install -r $HOME/app/requirements.txt
+CMD [ "python" , "/home/coronatracker/app/the_star.py"]
